linting ddx/pddx set
