--- applications/ddx/ddx.h
@@ -1,10 +1,0 @@
-/* ddx.h */
-
-double calculateRepulsion();
-double calculateEnergy(double test_diameter);
-
-void generateTestPoint();
-void findEnergyMinimum();
-void makeVerletList();
-void expandTestParticle();
-
