--- applications/ddx/io_setup.h
@@ -1,4 +1,0 @@
-/* io_setup.h */
-
-void readConfiguration();
-
