+++ applications/ddx/pddx.c
@@ -7,12 +7,53 @@
 #include <semaphore.h>
 #include <unistd.h>
 
-#include "ftw_param.h"
-#include "ftw_prng.h"
-#include "ftw_types.h"
-#include "ftw_std.h"
+#include <ftw_param.h>
+#include <ftw_prng.h>
+#include <ftw_types.h>
+#include <ftw_std.h>
+#include <io_setup.h>
 
-#include "pddx.h"
+/* was pddx.h */
+
+// setting defaults for ~640MB bss for 16M MOLECULES
+
+#ifndef MAX_NUM_MOLECULES
+#define MAX_NUM_MOLECULES 16777216
+#endif
+
+// setting defaults for 60MB of heap/thread for Verlet list of 1M
+
+#ifndef MAX_CLOSE
+#define MAX_CLOSE 1048576
+#endif
+
+typedef struct {
+  int				thread_id;
+  int   			close_molecules;
+  int   			attempts;
+  double 			test_x0, test_y0, test_z0;
+  double 			test_x, test_y, test_z;
+  double 			verlet_center_x, verlet_center_y, verlet_center_z;
+  double 			diameter;
+  double 			close_x[MAX_CLOSE], close_y[MAX_CLOSE], close_z[MAX_CLOSE];
+  double 			close_sigma[MAX_CLOSE];
+  double 			close_sigma6[MAX_CLOSE];
+  double 			close_sigma12[MAX_CLOSE];
+  double 			close_epsilon[MAX_CLOSE];
+  double 			sq_distance_from_initial_pt;
+  struct MersenneTwister 	rng;
+} Trajectory;
+
+double calculateRepulsion(Trajectory*);
+double calculateEnergy(Trajectory*, double);
+void generateTestPoint(Trajectory*);
+void findEnergyMinimum(Trajectory*);
+void makeVerletList(Trajectory*);
+void expandTestParticle(Trajectory*);
+void* ThreadMain(void *threadID);
+
+/* end of old pddx.h */
+
 
 pthread_t* 		threads;	// the threads
 void**			passvals;	// values passed to each thread
@@ -78,7 +119,7 @@ int main(int argc, char **argv)
     exit(0);
   }
 
-  readConfiguration();
+  loadConfiguration();
 
   // make and verify all the threads and resources
   passvals = (void **)malloc(sizeof(void*) * n_samples);
@@ -392,4 +433,5 @@ void expandTestParticle(Trajectory *p_traj)
   // copy diameter to trajectory data
   p_traj->diameter = diameter;
 }
+
